requirements.txt
opencv-python
numpy
mss
pytesseract
Pillow
pyinstaller